refactor: reorganize context menu hooks and improve menu closure handling in ContextMenu component
