+++ src/entities/context-menu/public/ContextMenu.tsx
@@ -4,18 +4,25 @@ import {
   useCallback,
   useRef,
   CSSProperties,
-  useState,
+  useMemo,
+  memo,
 } from "react";
 import { motion, AnimatePresence } from "framer-motion";
 import { IVector2D } from "@/lib/utils/data-structures/Vector2d";
 import buildClassName from "@/shared/lib/buildClassName";
 import stopEvent from "@/lib/utils/stopEvent";
-import useMenuPosition from "@/shared/hooks/DOM/useMenuPosition";
+import useMenuPosition from "@/entities/context-menu/public/hooks/useMenuPosition";
 
 import "./ContextMenu.scss";
-import { useClickAway } from "@/lib/hooks/history/events/useClick";
 import Portal from "@/shared/ui/Portal";
-import { useBoundaryCheck } from "@/shared/hooks/mouse/useBoundaryCheck";
+import useMenuClosure from "./hooks/useMenuClosure";
+
+const ANIMATION_PROPS = {
+  initial: { opacity: 0, scale: 0.85 },
+  animate: { opacity: 1, scale: 1 },
+  exit: { opacity: 0, scale: 0.85 },
+  transition: { duration: 0.15 },
+};
 
 interface ContextMenuProps {
   isOpen: boolean;
@@ -48,64 +55,59 @@ const ContextMenu: FC<ContextMenuProps> = ({
 }) => {
   const containerRef = useRef<HTMLDivElement>(null);
   const bubbleRef = useRef<HTMLDivElement>(null);
-  const [isHovered, setIsHovered] = useState(false);
-
-  useMenuPosition(isOpen, containerRef, bubbleRef, {
-    anchor: position,
-    getTriggerElement: () => rootRef?.current || document.body,
-    getRootElement: () => containerRef.current,
-    getMenuElement: () => bubbleRef.current,
-    getLayout: () => ({
-      isDense,
-      shouldAvoidNegativePosition: true,
-      withPortal: true,
-      menuElMinWidth: noCompact ? 220 : 160,
-    }),
-    withMaxHeight: true,
-  });
 
   const handleClose = useCallback(() => {
     onClose();
     document.removeEventListener("scroll", handleClose, true);
   }, [onClose]);
 
-  useClickAway(bubbleRef, handleClose);
+  const positionConfig = useMemo(
+    () => ({
+      anchor: position,
+      getTriggerElement: () => rootRef?.current || document.body,
+      getRootElement: () => containerRef.current,
+      getMenuElement: () => bubbleRef.current,
+      getLayout: () => ({
+        isDense,
+        shouldAvoidNegativePosition: true,
+        withPortal: true,
+        menuElMinWidth: noCompact ? 220 : 160,
+      }),
+      withMaxHeight: true,
+    }),
+    [position, isDense, noCompact, rootRef],
+  );
 
-  useBoundaryCheck({
-    elementRef: containerRef,
-    isActive: isOpen,
-    onExit: handleClose,
-    options: {
-      outboxSize: 100,
-      throttleInterval: 100,
-    },
-  });
+  useMenuPosition(isOpen, containerRef, bubbleRef, positionConfig);
+  useMenuClosure(isOpen, containerRef, handleClose);
 
-  const menuEl = (
-    <motion.div
-      ref={containerRef}
-      className={buildClassName("context-menu-container", className)}
-      style={style}
-      initial={{ opacity: 0, scale: 0.85 }}
-      animate={{ opacity: 1, scale: 1 }}
-      exit={{ opacity: 0, scale: 0.85 }}
-      transition={{ duration: 0.2 }}
-      onContextMenu={stopEvent}
-    >
-      <div
-        className={buildClassName("context-menu-bubble", menuClassName)}
-        ref={bubbleRef}
+  const menuEl = useMemo(
+    () => (
+      <motion.div
+        key={"context-menu"}
+        ref={containerRef}
+        className={buildClassName("context-menu-container", className)}
+        style={style}
+        {...ANIMATION_PROPS}
+        onContextMenu={stopEvent}
+        onAnimationEnd={onCloseAnimationEnd}
       >
-        {children}
-      </div>
-    </motion.div>
+        <div
+          className={buildClassName("context-menu-bubble", menuClassName)}
+          ref={bubbleRef}
+        >
+          {children}
+        </div>
+      </motion.div>
+    ),
+    [className, menuClassName, style, children, onCloseAnimationEnd],
   );
 
   return (
-    <AnimatePresence>
+    <AnimatePresence mode="wait">
       {isOpen && (withPortal ? <Portal>{menuEl}</Portal> : menuEl)}
     </AnimatePresence>
   );
 };
 
-export default ContextMenu;
+export default memo(ContextMenu);
